refactor: 슬라이스 정리 (#ATR-410)
