+++ apps/service/entities/theme/index.ts
@@ -1,3 +1,4 @@
 export * from './lib'
 export * from './model'
 export * from './constant'
+export * from './ui'
