--- apps/service/features/set-theme/index.ts
@@ -1,1 +1,0 @@
-export * from './ui'
